Update arraycontext/loopy.py, arraycontext/fake_numpy.py, arraycontext/impl/pyopencl/fake_numpy.py files
